renamed appointment.ts and review.ts to appointments.ts and reviews.ts respectively, for consistency
