+++ typescriptSample/src/data/appointments.ts
